requirements.txt
langchain
langchain-groq
langchain-google-genai
langchain-community
gradio
tavily-python
python-dotenv
